책 정보 자동 저장 기능추가(파싱 후 디비까지 저장하는 로직), 테스트코드

+++ domain/src/main/java/sion/bookstore/domain/book/repository/BookRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 @Mapper
 @Repository
 public interface BookRepository {
-    Integer create(Book book);
+    Long create(Book book);
     Book findOne(Long bookId);
     void update(Book book);
 //    List<Book> findAll(BookSearchCondition bookSearchCondition);
